requirements.txt
pygithub==1.55.0
tweepy==4.5.0
